moves modules in to the ledcontroller folder
